refactor: clean up core types and fix tl::expected deprecation
Extract estimator structs into structs.hpp and add make_not_updated() to
reduce repeated failure handling. Switch from deprecated tl_expected to
rcpputils/tl_expected, narrow rclcpp includes, and add [[nodiscard]].
